# main_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Usta Top - vakansiya va resume platformasi API",
        "title": "Usta Top API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@usta-top.uz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/vacancies": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Vakansiyalar ro'yxati",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nomi",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sarlavha",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Foydalanuvchi ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Region ID",
                        "name": "region_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Tuman ID",
                        "name": "district_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Mahalla ID",
                        "name": "mahalla_id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Faol/faolsiz",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimal narx",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maksimal narx",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Sahifa",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Saralash maydoni",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc yoki desc",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Yangi vakansiya yaratish",
                "parameters": [
                    {
                        "description": "Vakansiya ma'lumotlari",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vacancy_dto.CreateVacancyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/vacancy_dto.VacancyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vacancies/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Vakansiyani yangilash",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vakansiya ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Yangi ma'lumotlar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vacancy_dto.UpdateVacancyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vacancy_dto.VacancyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Vakansiyani o'chirish",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vakansiya ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/vacancies/{slug}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Vakansiyani slug bo'yicha olish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Vakansiya slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vacancy_dto.VacancyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "vacancy_dto.CreateVacancyRequest": {
            "type": "object",
            "required": [
                "adress",
                "contact",
                "name",
                "text",
                "title"
            ],
            "properties": {
                "adress": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "contact": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                },
                "district_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "mahalla_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "price": {
                    "type": "integer"
                },
                "region_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string",
                    "minLength": 10
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 2
                }
            }
        },
        "vacancy_dto.UpdateVacancyRequest": {
            "type": "object",
            "properties": {
                "adress": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "contact": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 5
                },
                "district_id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "mahalla_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 2
                },
                "price": {
                    "type": "integer"
                },
                "region_id": {
                    "type": "integer"
                },
                "text": {
                    "type": "string",
                    "minLength": 10
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 2
                }
            }
        },
        "vacancy_dto.VacancyResponse": {
            "type": "object",
            "properties": {
                "adress": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "district_id": {
                    "type": "integer"
                },
                "district_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "mahalla_id": {
                    "type": "integer"
                },
                "mahalla_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "region_id": {
                    "type": "integer"
                },
                "region_name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "views_count": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}